# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Sales Platform API
  version: 2.0.0
  description: Authenticated, tenant-isolated CRM, reporting, settings, and AI workflows.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /operations/inbox:
    get:
      operationId: getHumanReplyInbox
      tags: [Operations]
      summary: List owned replies requiring human judgment
      responses:
        "200": { description: Reply inbox returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /operations/tasks:
    get:
      operationId: listTasks
      tags: [Operations]
      summary: List owned operational tasks
      responses:
        "200": { description: Tasks returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /operations/analytics:
    get:
      operationId: getCampaignAnalytics
      tags: [Operations]
      summary: Return metrics computed from stored events only
      responses:
        "200": { description: Observed analytics returned }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: refresh_token }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
...
  responses:
...
    Unauthorized: { description: Authentication failed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...